+++ toucan_connectors/mssql_TLSv1_0/mssql_connector.py
@@ -108,7 +108,7 @@ class MSSQLConnector(ToucanConnector, data_source_model=MSSQLDataSource):
         if self.port is not None:
             server += f",{self.port}"
         con_params = {
-            "driver": "{ODBC Driver 17 for SQL Server}",
+            "driver": "{ODBC Driver 18 for SQL Server}",
             "server": server,
             "database": database,
             "user": self.user,